backend/requirements.txt
# ──────────────────────────────────────────────
# Expense Sense – Backend Dependencies
# ──────────────────────────────────────────────

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# MongoDB async driver
motor==3.6.0
pymongo>=4.9,<4.10

# Authentication & Security
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==44.0.0

# Data validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Machine Learning
scikit-learn==1.6.0
numpy==2.2.1
joblib==1.4.2

# Environment
python-dotenv==1.0.1

# Dev / Testing
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
